De-inline Socket.hpp into Socket.cpp
Maybe this could be part of #176 ? I think it also, on some compilers,
reveals some warnings about using strncpy and memcmp with sockaddr
stuff.
