+++ src/Makefile.in
@@ -36,6 +36,7 @@ LIBBIFROST_OBJS = \
   testsuite.o \
   ring.o \
   ring_impl.o \
+  Socket.o \
   array.o \
   unpack.o \
   quantize.o \
